requirements.txt
robonomics-interface==1.6.1
discord.py==2.3.2
pinatapy-vourhey==0.1.3
